fix(articles): re-implement missing call to mark articles as visited
The previous refactor to separate article metadata from content
in commit bf5bc75878eb09747fb31cf0018ca9e1e702cf56 accidentally
removed the call to `markAsVisited` when an article is viewed.

This change re-introduces the functionality by creating a new
`ArticleViewTracker` component that uses the `useVisitedArticles`
hook to mark articles as visited when the article page is loaded.

This resolves the bug where the number of unseen articles was not
updating correctly.